feat: delete recipients instead of anonymizing them
