+++ .ci/appsettings.override.postgres.local.json
@@ -21,9 +21,6 @@
     },
     "Modules": {
         "Announcements": {
-            "Application": {
-                "DidDomainName": "localhost"
-            },
             "Infrastructure": {
                 "SqlDatabase": {
                     "Provider": "Postgres",
