+++ .ci/appsettings.override.sqlserver.docker.json
@@ -21,9 +21,6 @@
     },
     "Modules": {
         "Announcements": {
-            "Application": {
-                "DidDomainName": "localhost"
-            },
             "Infrastructure": {
                 "SqlDatabase": {
                     "Provider": "SqlServer",
