+++ Applications/AdminApi/test/AdminApi.Tests.Integration/api.appsettings.local.override.json
@@ -11,9 +11,6 @@
     },
     "Modules": {
         "Announcements": {
-            "Application": {
-                "DidDomainName": "localhost"
-            },
             "Infrastructure": {
                 "SqlDatabase": {
                     "Provider": "Postgres",
