+++ Applications/IdentityDeletionJobs/src/Job.IdentityDeletion/ServicesExtensions.cs
@@ -6,6 +6,7 @@ public static class ServicesExtensions
 {
     public static IServiceCollection RegisterIdentityDeleters(this IServiceCollection services)
     {
+        services.AddTransient<IIdentityDeleter, Modules.Announcements.Application.Announcements.IdentityDeleter>();
         services.AddTransient<IIdentityDeleter, Modules.Challenges.Application.Identities.IdentityDeleter>();
         services.AddTransient<IIdentityDeleter, Modules.Devices.Application.Identities.IdentityDeleter>();
         services.AddTransient<IIdentityDeleter, Modules.Files.Application.Identities.IdentityDeleter>();
@@ -14,7 +15,6 @@ public static class ServicesExtensions
         services.AddTransient<IIdentityDeleter, Modules.Relationships.Application.Identities.IdentityDeleter>();
         services.AddTransient<IIdentityDeleter, Modules.Synchronization.Application.Identities.IdentityDeleter>();
         services.AddTransient<IIdentityDeleter, Modules.Tokens.Application.Identities.IdentityDeleter>();
-        services.AddTransient<IIdentityDeleter, Modules.Announcements.Application.Announcements.IdentityDeleter>();
 
         return services;
     }
