--- Modules/Announcements/src/Announcements.Application/Announcements/Commands/AnonymizeRecipient/AnonymizeRecipientForIdentityCommand.cs
@@ -1,13 +1,0 @@
-﻿using MediatR;
-
-namespace Backbone.Modules.Announcements.Application.Announcements.Commands.AnonymizeRecipient;
-
-public record AnonymizeRecipientForIdentityCommand : IRequest
-{
-    public AnonymizeRecipientForIdentityCommand(string identityAddress)
-    {
-        IdentityAddress = identityAddress;
-    }
-
-    public string IdentityAddress { get; }
-}
