+++ Modules/Announcements/src/Announcements.Application/Announcements/Commands/DeleteAnnouncementRecipients/DeleteAnnouncementRecipientsCommand.cs
@@ -1,0 +1,13 @@
+﻿using MediatR;
+
+namespace Backbone.Modules.Announcements.Application.Announcements.Commands.DeleteAnnouncementRecipients;
+
+public record DeleteAnnouncementRecipientsCommand : IRequest
+{
+    public DeleteAnnouncementRecipientsCommand(string identityAddress)
+    {
+        IdentityAddress = identityAddress;
+    }
+
+    public string IdentityAddress { get; }
+}
