+++ Modules/Announcements/src/Announcements.Application/Announcements/IdentityDeleter.cs
@@ -1,6 +1,6 @@
 ﻿using Backbone.BuildingBlocks.Application.Identities;
 using Backbone.DevelopmentKit.Identity.ValueObjects;
-using Backbone.Modules.Announcements.Application.Announcements.Commands.AnonymizeRecipient;
+using Backbone.Modules.Announcements.Application.Announcements.Commands.DeleteAnnouncementRecipients;
 using MediatR;
 
 namespace Backbone.Modules.Announcements.Application.Announcements;
@@ -18,7 +18,7 @@ public class IdentityDeleter : IIdentityDeleter
 
     public async Task Delete(IdentityAddress identityAddress)
     {
-        await _mediator.Send(new AnonymizeRecipientForIdentityCommand(identityAddress));
+        await _mediator.Send(new DeleteAnnouncementRecipientsCommand(identityAddress));
         await _deletionProcessLogger.LogDeletion(identityAddress, "AnnouncementRecipients");
     }
 }
