+++ Modules/Announcements/src/Announcements.Application/Infrastructure/Persistence/Repository/IAnnouncementsRepository.cs
@@ -7,6 +7,6 @@ public interface IAnnouncementsRepository
 {
     Task Add(Announcement announcement, CancellationToken cancellationToken);
     Task<List<Announcement>> FindAll(CancellationToken cancellationToken);
-    Task<List<Announcement>> FindAllForIdentityAddress(Expression<Func<Announcement, bool>> filter, CancellationToken cancellationToken);
     Task Update(List<Announcement> announcements, CancellationToken cancellationToken);
+    Task DeleteRecipients(Expression<Func<AnnouncementRecipient, bool>> filter, CancellationToken cancellationToken);
 }
