+++ appsettings.override.json
@@ -33,9 +33,6 @@
     },
     "Modules": {
         "Announcements": {
-            "Application": {
-                "DidDomainName": "localhost"
-            },
             "Infrastructure": {
                 "SqlDatabase": {
                     "Provider": "Postgres",
